requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
Flask-JWT==0.3.2 
mysqlclient==1.4.6
mongoengine==0.23.1 
werkzeug==0.16.0
uWSGI==2.0.19.1